Refactor to keep node-mime specific enhancements out of the Apache types file.
The _init() method is now mime.load(), and can be called more than once to load
multiple  files.  This allows us to break node-mime definitions into:

- mime.types:  The Apache mime.types file
- node.types:  node-mime specific enhancements

The rational for this is that we want to avoid making edits to this file
locally because it complicates the process of updating the file when Apache
updates their version.  So instead, use node.types to define any enhancements
that we can then "overlay" on top of whatever is defined in mime.types.