Bind global methods
This module exports methods on an instance of the `Mime` class, meaning
they must be invoked in the context of the module itself. It is common
for consumer code to load only specific methods of a given module, as
in:

    var extensionFromMime = require('mime').extension;

In order to facilitate this pattern, the exported instance methods
should be bound to the instance.